[fix]: a careless mistake...

+++ Programming/Nevo/Nevo/Class/Controller/LoginController/LoginController.swift
@@ -215,7 +215,7 @@ class LoginController: UIViewController,UITextFieldDelegate {
                         let profile:UserProfile = user!
                         let dateString:String = Date(timeIntervalSince1970: stepsModel.date).stringFromFormat("yyy-MM-dd")
                         var caloriesValue:Int = 0
-                        var milesValue:Int = 0
+                        var milesValue:Double = 0
                         StepGoalSetingController.calculationData((stepsModel.walking_duration+stepsModel.running_duration), steps: stepsModel.steps, completionData: { (miles, calories) in
                             caloriesValue = Int(calories)
                             milesValue = miles
